Changes on keyboard press are incremental.
The previous implementation naively set the value of the force of the
appropriate thrusters.

Now, a subscriber is declared inside the same node. And this subscribes
to the present value, and when a new change has to be done, the value of
the force is either increased or decreased from the earlier value.